requirements.txt
pyyaml
tqdm
python-pptx==0.6.18
